Move ST_Explode from table to properties pkg
